Modified `option` struct so that it no longer throws errors, was using floats where I didn't need to... - Moved all test files into `test` - Some other small changes to accompany the above changes
