+++ test/instrument_test.go
@@ -1,17 +1,19 @@
-package instrument
+package test
 
 import (
 	"testing"
+
+	"github.com/samjtro/go-tda/instrument"
 )
 
 func TestInstrument(t *testing.T) {
-	_, err := Simple("AAPL")
+	_, err := instrument.Simple("AAPL")
 
 	if err != nil {
 		t.Fatalf(err.Error())
 	}
 
-	_, err = Fundamental("AAPL")
+	_, err = instrument.Fundamental("AAPL")
 
 	if err != nil {
 		t.Fatalf(err.Error())
